+++ PKHeX.Core/Editing/Bulk/Suggestion/BatchModifications.cs
@@ -22,7 +22,7 @@ internal static class BatchModifications
             }
             else if (!IsNone(propValue))
             {
-                Span<int> moves = stackalloc int[4];
+                Span<ushort> moves = stackalloc ushort[4];
                 pk.GetMoves(moves);
                 t.SetRecordFlags(moves); // whatever fit the current moves
             }
@@ -90,7 +90,7 @@ internal static class BatchModifications
     /// </summary>
     /// <param name="pk">Pokémon to modify.</param>
     /// <param name="moves">Moves to apply.</param>
-    public static ModifyResult SetMoves(PKM pk, ReadOnlySpan<int> moves)
+    public static ModifyResult SetMoves(PKM pk, ReadOnlySpan<ushort> moves)
     {
         pk.SetMoves(moves);
         pk.HealPP();
