+++ PKHeX.Core/Legality/LearnSource/Group/ILearnGroup.cs
@@ -17,7 +17,7 @@ public interface ILearnGroup
     /// </summary>
     bool HasVisited(PKM pk, EvolutionHistory history);
 
-    bool Check(Span<MoveResult> result, ReadOnlySpan<int> current, PKM pk, EvolutionHistory history, IEncounterTemplate enc, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current);
+    bool Check(Span<MoveResult> result, ReadOnlySpan<ushort> current, PKM pk, EvolutionHistory history, IEncounterTemplate enc, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current);
 
     void GetAllMoves(Span<bool> result, PKM pk, EvolutionHistory history, IEncounterTemplate enc, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current);
 }
