+++ PKHeX.Core/Legality/LearnSource/Group/LearnGroup7b.cs
@@ -13,7 +13,7 @@ public sealed class LearnGroup7b : ILearnGroup
     public ILearnGroup? GetPrevious(PKM pk, EvolutionHistory history, IEncounterTemplate enc, LearnOption option) => null;
     public bool HasVisited(PKM pk, EvolutionHistory history) => history.HasVisitedLGPE;
 
-    public bool Check(Span<MoveResult> result, ReadOnlySpan<int> current, PKM pk, EvolutionHistory history,
+    public bool Check(Span<MoveResult> result, ReadOnlySpan<ushort> current, PKM pk, EvolutionHistory history,
         IEncounterTemplate enc, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current)
     {
         var evos = history.Gen7b;
@@ -23,7 +23,7 @@ public sealed class LearnGroup7b : ILearnGroup
         return MoveResult.AllParsed(result);
     }
 
-    private static void Check(Span<MoveResult> result, ReadOnlySpan<int> current, PKM pk, EvoCriteria evo, int stage)
+    private static void Check(Span<MoveResult> result, ReadOnlySpan<ushort> current, PKM pk, EvoCriteria evo, int stage)
     {
         var game = LearnSource7GG.Instance;
         if (!game.TryGetPersonal(evo.Species, evo.Form, out var pi))
