+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource3LG.cs
@@ -32,7 +32,7 @@ public sealed class LearnSource3LG : ILearnSource, IEggSource
         return true;
     }
 
-    public bool GetIsEggMove(int species, int form, int move)
+    public bool GetIsEggMove(int species, int form, ushort move)
     {
         if ((uint)species > MaxSpecies)
             return false;
@@ -40,14 +40,14 @@ public sealed class LearnSource3LG : ILearnSource, IEggSource
         return moves.GetHasEggMove(move);
     }
 
-    public ReadOnlySpan<int> GetEggMoves(int species, int form)
+    public ReadOnlySpan<ushort> GetEggMoves(int species, int form)
     {
         if ((uint)species > MaxSpecies)
-            return ReadOnlySpan<int>.Empty;
+            return ReadOnlySpan<ushort>.Empty;
         return EggMoves[species].Moves;
     }
 
-    public MoveLearnInfo GetCanLearn(PKM pk, PersonalInfo pi, EvoCriteria evo, int move, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current)
+    public MoveLearnInfo GetCanLearn(PKM pk, PersonalInfo pi, EvoCriteria evo, ushort move, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current)
     {
         if (types.HasFlagFast(MoveSourceType.LevelUp))
         {
@@ -71,7 +71,7 @@ public sealed class LearnSource3LG : ILearnSource, IEggSource
         return default;
     }
 
-    private static bool GetIsTutor(int species, int move) => move switch
+    private static bool GetIsTutor(int species, ushort move) => move switch
     {
         (int)Move.BlastBurn => species == (int)Species.Charizard,
         (int)Move.HydroCannon => species == (int)Species.Blastoise,
@@ -79,7 +79,7 @@ public sealed class LearnSource3LG : ILearnSource, IEggSource
         _ => false,
     };
 
-    private static bool GetIsTM(PersonalInfo info, int move)
+    private static bool GetIsTM(PersonalInfo info, ushort move)
     {
         var index = Array.IndexOf(TM_3, move);
         if (index == -1)
@@ -87,7 +87,7 @@ public sealed class LearnSource3LG : ILearnSource, IEggSource
         return info.TMHM[index];
     }
 
-    private static bool GetIsHM(PersonalInfo info, int move)
+    private static bool GetIsHM(PersonalInfo info, ushort move)
     {
         var index = Array.IndexOf(HM_3, move);
         if (index == -1)
