+++ PKHeX.Core/Legality/LearnSource/Sources/Shared/ILearnSource.cs
@@ -18,7 +18,7 @@ public interface ILearnSource
     /// <param name="types">Types of move sources to iterate</param>
     /// <param name="option">Option to check if it can be currently known, or previously known.</param>
     /// <returns>Details about how the move can be learned. Will be equivalent to default if cannot learn.</returns>
-    public MoveLearnInfo GetCanLearn(PKM pk, PersonalInfo pi, EvoCriteria evo, int move, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current);
+    public MoveLearnInfo GetCanLearn(PKM pk, PersonalInfo pi, EvoCriteria evo, ushort move, MoveSourceType types = MoveSourceType.All, LearnOption option = LearnOption.Current);
 
     /// <summary>
     /// Yields an iterable list of all potential moves that an <see cref="evo"/> can learn from this <see cref="ILearnSource"/>.
